+++ default_tasks/clean.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/bash
 
-delete_dir "$OBJDIR"
-delete_dir "$OUTDIR"
+try_delete_dir_or_file "$OBJDIR"
+try_delete_dir_or_file "$OUTDIR"
 myprint "${WHITE}deleting build logs"
 rm -rf *.log